rework file structure to fix imports from rasa.shared
