+++ rasa/nlu/classifiers/llm_flow_classifier.py
@@ -42,7 +42,7 @@ from rasa.shared.constants import (
 )
 from rasa.shared.nlu.training_data.message import Message
 from rasa.shared.nlu.training_data.training_data import TrainingData
-from rasa.utils.llm import (
+from rasa.shared.utils.llm import (
     DEFAULT_OPENAI_CHAT_MODEL_NAME,
     llm_factory,
     tracker_as_readable_transcript,
